♻️ refactor/KIKI-61 : ELK 관련 도커 수정
