+++ .github/workflows/dev-ci-cd.yml
@@ -9,7 +9,7 @@ name: 키키하이 dev CI-CD 파이프라인
 
 on:
   push:
-    branches: [ "develop" ]
+    branches: [ "develop","feat/product/KIKI-61-ElasticSearch-Dev" ]
 
 jobs:
   #1. 개발 서버 CI, Build 용
