+++ docker/logstash/config/logstash.yml
@@ -1,0 +1,16 @@
+# 기본 설정
+http.host: "0.0.0.0"
+xpack.monitoring.enabled: false  # Elasticsearch 모니터링을 사용하지 않을 경우 false
+log.level: info
+
+# 파이프라인 경로 (기본적으로 이 경로를 봄)
+path.config: /usr/share/logstash/pipeline
+
+# 로그 파일 경로
+path.logs: /usr/share/logstash/logs
+
+# 데이터 경로
+path.data: /usr/share/logstash/data
+
+config.reload.automatic: true  # 설정 파일 변경 시 자동으로 재시작
+config.reload.interval: 3s  # 설정 파일 변경 감지 주기
