+++ docker/logstash/pipeline/logstash.conf
@@ -1,0 +1,25 @@
+input {
+  http {
+    port => 5044
+  }
+    mongodb {
+      uri => "${MONGO_ATLAS}"
+      placeholder_db_dir => '/opt/logstash-mongodb/'
+      placeholder_db_name => 'logstash_sqlite.db'
+      collection => 'products'
+      batch_size => 1000
+    }
+}
+
+
+output {
+  elasticsearch {
+    hosts => ["https://es01:9200"]
+    user => "elastic"
+    password => "${ELASTIC_PASSWORD}"
+    cacert => "/usr/share/logstash/certs/ca/ca.crt"
+    index => "products"
+    action => "index"
+  }
+  stdout { codec => rubydebug }
+}
