+++ src/main/java/com/jiyoung/kikihi/global/config/external/ElasticSearchConfig.java
@@ -1,0 +1,26 @@
+package com.jiyoung.kikihi.global.config.external;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.data.elasticsearch.client.ClientConfiguration;
+import org.springframework.data.elasticsearch.client.elc.ElasticsearchConfiguration;
+import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories;
+
+@Configuration
+@Profile("dev")
+@EnableElasticsearchRepositories(basePackages = "com.jiyoung.kikihi.platform.adapter.out.elasticSearch")
+public class ElasticSearchConfig extends ElasticsearchConfiguration {
+
+    @Value("${spring.elasticsearch.uris}")
+    private String host;
+
+    @Override
+    public ClientConfiguration clientConfiguration() {
+        return ClientConfiguration.builder()
+                .connectedTo(host)
+                .usingSsl(false)
+                .build();
+    }
+
+}
